Fix the docs
