+++ docs/api/sql/Raster-operators.md
@@ -7,7 +7,7 @@ Format: `RS_AddBand (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val sumDF = spark.sql("select RS_Add(band1, band2) as sumOfBands from dataframe")
 
@@ -22,7 +22,7 @@ Format: `RS_SubtractBands (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val subtractDF = spark.sql("select RS_Subtract(band1, band2) as differenceOfOfBands from dataframe")
 
@@ -37,7 +37,7 @@ Format: `RS_MultiplyBands (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val multiplyDF = spark.sql("select RS_MultiplyBands(band1, band2) as multiplyBands from dataframe")
 
@@ -52,7 +52,7 @@ Format: `RS_DivideBands (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val multiplyDF = spark.sql("select RS_DivideBands(band1, band2) as divideBands from dataframe")
 
@@ -67,7 +67,7 @@ Format: `RS_MultiplyFactor (Band1: Array[Double], Factor: Int)`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val multiplyFactorDF = spark.sql("select RS_MultiplyFactor(band1, 2) as multiplyfactor from dataframe")
 
@@ -82,7 +82,7 @@ Format: `RS_Mode (Band: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val modeDF = spark.sql("select RS_Mode(band) as mode from dataframe")
 
@@ -97,7 +97,7 @@ Format: `RS_Mean (Band: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val meanDF = spark.sql("select RS_Mean(band) as mean from dataframe")
 
@@ -112,7 +112,7 @@ Format: `RS_NormalizedDifference (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val normalizedDF = spark.sql("select RS_NormalizedDifference(band1, band2) as normdifference from dataframe")
 
@@ -127,7 +127,7 @@ Format: `RS_Count (Band1: Array[Double], Target: Double)`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val countDF = spark.sql("select RS_Count(band1, target) as count from dataframe")
 
@@ -142,7 +142,7 @@ Format: `RS_GreaterThan (Band: Array[Double], Target: Double)`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val greaterDF = spark.sql("select RS_GreaterThan(band, target) as maskedvalues from dataframe")
 
@@ -157,7 +157,7 @@ Format: `RS_GreaterThanEqual (Band: Array[Double], Target: Double)`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val greaterEqualDF = spark.sql("select RS_GreaterThanEqual(band, target) as maskedvalues from dataframe")
 
@@ -172,7 +172,7 @@ Format: `RS_LessThan (Band: Array[Double], Target: Double)`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val lessDF = spark.sql("select RS_LessThan(band, target) as maskedvalues from dataframe")
 
@@ -187,7 +187,7 @@ Format: `RS_LessThanEqual (Band: Array[Double], Target: Double)`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val lessEqualDF = spark.sql("select RS_LessThanEqual(band, target) as maskedvalues from dataframe")
 
@@ -202,7 +202,7 @@ Format: `RS_Modulo (Band: Array[Double], Target: Double)`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val moduloDF = spark.sql("select RS_Modulo(band, target) as modulo from dataframe")
 
@@ -217,7 +217,7 @@ Format: `RS_BitwiseAND (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val biwiseandDF = spark.sql("select RS_BitwiseAND(band1, band2) as andvalue from dataframe")
 
@@ -232,7 +232,7 @@ Format: `RS_BitwiseOR (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val biwiseorDF = spark.sql("select RS_BitwiseOR(band1, band2) as or from dataframe")
 
@@ -247,7 +247,7 @@ Format: `RS_SquareRoot (Band: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val rootDF = spark.sql("select RS_SquareRoot(band) as squareroot from dataframe")
 
@@ -262,7 +262,7 @@ Format: `RS_LogicalDifference (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val logicalDifference = spark.sql("select RS_LogicalDifference(band1, band2) as logdifference from dataframe")
 
@@ -277,7 +277,7 @@ Format: `RS_LogicalOver (Band1: Array[Double], Band2: Array[Double])`
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val logicalOver = spark.sql("select RS_LogicalOver(band1, band2) as logover from dataframe")
 
@@ -292,8 +292,18 @@ Format: `RS_FetchRegion (Band: Array[Double], coordinates: Array[Int], dimenstio
 Since: `v1.1.0`
 
 Spark SQL example:
-```SQL
+```Scala
 
 val region = spark.sql("select RS_FetchRegion(Band,Array(0, 0, 1, 2),Array(3, 3)) as Region from dataframe")
+```
 
+## RS_Normalize
+
+Introduction: Normalize the value in the array to [0, 255]
+
+Since: `v1.1.0`
+
+Spark SQL example
+```SQL
+SELECT RS_Normalize(band)
 ```
