+++ sql/src/main/scala/org/apache/spark/sql/sedona_sql/expressions/raster/Functions.scala
@@ -824,6 +824,42 @@ case class RS_LogicalOver(inputExpressions: Seq[Expression])
   override def children: Seq[Expression] = inputExpressions
 }
 
+case class RS_Normalize(inputExpressions: Seq[Expression])
+  extends Expression with CodegenFallback with UserDataGeneratator {
+  override def nullable: Boolean = false
 
+  override def eval(inputRow: InternalRow): Any = {
+    // This is an expression which takes one input expressions
+    var band:Array[Double] = null
+
+    if(inputExpressions(0).eval(inputRow).getClass.toString() == "class org.apache.spark.sql.catalyst.expressions.UnsafeArrayData") {
+      band =inputExpressions(0).eval(inputRow).asInstanceOf[UnsafeArrayData].toDoubleArray()
+    }
+    else {
+      band =inputExpressions(0).eval(inputRow).asInstanceOf[GenericArrayData].toDoubleArray()
+
+    }
+    val result = normalize(band)
+    new GenericArrayData(result)
+  }
+
+  // Normalize between 0 and 255
+  private def normalize(band: Array[Double]): Array[Double] = {
+
+    val result = new Array[Double](band.length)
+    val maxVal = band.toList.max
+
+    for(i<-0 until band.length) {
+      result(i) = (band(i)/(maxVal/255.0)).toInt
+    }
+
+    result
+
+  }
+
+  override def dataType: DataType = ArrayType(DoubleType)
+
+  override def children: Seq[Expression] = inputExpressions
+}
